requirements.txt
aiohttp==3.8.1; python_version >= "3.6"
aiosignal==1.2.0; python_version >= "3.6"
anyio==3.5.0; python_full_version >= "3.6.2" and python_version >= "3.6"
async-timeout==4.0.2; python_version >= "3.6"
attrs==21.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
certifi==2021.10.8; python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.5.0" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.5" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.5" and python_full_version >= "3.5.0"
cssselect==1.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
frozenlist==1.3.0; python_version >= "3.7"
h11==0.12.0; python_version >= "3.6"
h2==4.1.0; python_version >= "3.6" and python_full_version >= "3.6.1"
hpack==4.0.0; python_version >= "3.6" and python_full_version >= "3.6.1"
httpcore==0.14.7; python_version >= "3.6"
httpx==0.22.0; python_version >= "3.6"
hyperframe==6.0.1; python_version >= "3.6" and python_full_version >= "3.6.1"
idna==3.3; python_version >= "3.6" and python_full_version >= "3.6.2"
loguru==0.6.0; python_version >= "3.5"
lxml==4.8.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
multidict==6.0.2; python_version >= "3.7"
parsel==1.6.0
pyyaml==6.0; python_version >= "3.6"
rfc3986==1.5.0; python_version >= "3.6"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sniffio==1.2.0; python_full_version >= "3.6.2" and python_version >= "3.6"
telepot==12.7
urllib3==1.26.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
w3lib==1.22.0
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
yarl==1.7.2; python_version >= "3.6"
